backend/requirements.txt
httpx
requests
fastapi
uvicorn
pydantic
python-multipart
pypdf2
pymupdf
python-dotenv
ollama
pydantic-settings
reportlab

